Enum libp2p_kad::KademliaOut

pub enum KademliaOut {
    Discovered {
        peer_id: PeerId,
        addresses: Vec<Multiaddr>,
        ty: KadConnectionType,
    },
    KBucketAdded {
        peer_id: PeerId,
        replaced: Option<PeerId>,
    },
    FindNodeResult {
        key: PeerId,
        closer_peers: Vec<PeerId>,
    },
    GetProvidersResult {
        key: Multihash,
        provider_peers: Vec<PeerId>,
        closer_peers: Vec<PeerId>,
    },
    GetValueResult(GetValueResult),
    PutValueResult(PutValueResult),
}

Output event of the Kademlia behaviour.

Variants
Discovered

We have discovered a node.

Note: The Kademlia behaviour doesn't store the addresses of this node, and therefore attempting to connect to this node may or may not work.

Fields of Discovered

peer_id: PeerId

Id of the node that was discovered.

addresses: Vec<Multiaddr>

Addresses of the node.

ty: KadConnectionType

How the reporter is connected to the reported.

KBucketAdded

A node has been added to a k-bucket.

Fields of KBucketAdded

peer_id: PeerId

Id of the node that was added.

replaced: Option<PeerId>

If Some, this addition replaced the value that is inside the option.

FindNodeResult

Result of a FIND_NODE iterative query.

Fields of FindNodeResult

key: PeerId

The key that we looked for in the query.

closer_peers: Vec<PeerId>

List of peers ordered from closest to furthest away.

GetProvidersResult

Result of a GET_PROVIDERS iterative query.

Fields of GetProvidersResult

key: Multihash

The key that we looked for in the query.

provider_peers: Vec<PeerId>

The peers that are providing the requested key.

closer_peers: Vec<PeerId>

List of peers ordered from closest to furthest away.

GetValueResult(GetValueResult)

Result of a GET_VALUE query

PutValueResult(PutValueResult)

Result of a PUT_VALUE query