Module libp2p_kad::handler

Structs

KademliaHandler

Protocol handler that manages substreams for the Kademlia protocol on a single connection with a peer.

KademliaHandlerConfig

Configuration of a KademliaHandler.

KademliaHandlerProto

A prototype from which KademliaHandlers can be constructed.

KademliaRequestId

Unique identifier for a request. Must be passed back in order to answer a request from the remote.

Enums

KademliaHandlerEvent

Event produced by the Kademlia handler.

KademliaHandlerIn

Event to send to the handler.

KademliaHandlerQueryErr

Error that can happen when requesting an RPC query.