Struct libp2p_kad::handler::KademliaHandler

pub struct KademliaHandler<TSubstream, TUserData> where
    TSubstream: AsyncRead + AsyncWrite,  { /* fields omitted */ }

Protocol handler that handles Kademlia communications with the remote.

The handler will automatically open a Kademlia substream with the remote for each request we make.

It also handles requests made by the remote.

Methods

impl<TSubstream, TUserData> KademliaHandler<TSubstream, TUserData> where
TSubstream: AsyncRead + AsyncWrite,

pub fn dial_only() -> Self

Create a KademliaHandler that only allows sending messages to the remote but denying incoming connections.

pub fn dial_and_listen() -> Self

Create a KademliaHandler that only allows sending messages but also receive incoming requests.

The Default trait implementation wraps around this function.

pub fn with_protocol_name(self, name: impl Into<Cow<'static, [u8]>>) -> Self

Modifies the protocol name used on the wire. Can be used to create incompatibilities between networks on purpose.