Struct libp2p_kad::kbucket::KBucketRef

pub struct KBucketRef<'a, TKey, TVal> { /* private fields */ }

A reference to a bucket in a KBucketsTable.

Implementations

impl<'a, TKey, TVal> KBucketRef<'a, TKey, TVal> where
TKey: Clone + AsRef<KeyBytes>,
TVal: Clone,

pub fn range(&self) -> (Distance, Distance)

Returns the minimum inclusive and maximum inclusive Distance for this bucket.

pub fn is_empty(&self) -> bool

Checks whether the bucket is empty.

pub fn num_entries(&self) -> usize

Returns the number of entries in the bucket.

pub fn has_pending(&self) -> bool

Returns true if the bucket has a pending node.

pub fn contains(&self, d: &Distance) -> bool

Tests whether the given distance falls into this bucket.

pub fn rand_distance(&self, rng: &mut impl Rng) -> Distance

Generates a random distance that falls into this bucket.

Together with a known key a (e.g. the local key), a random distance d for this bucket w.r.t k gives rise to the corresponding (random) key b s.t. the XOR distance between a and b is d. In other words, it gives rise to a random key falling into this bucket. See key::Key::for_distance.

pub fn iter(&'a self) -> impl Iterator<Item = EntryRefView<'a, TKey, TVal>>

Returns an iterator over the entries in the bucket.