Struct libp2p_kad::kbucket::KBucketsTable

pub struct KBucketsTable<TPeerId, TVal> { /* fields omitted */ }

A KBucketsTable represents a Kademlia routing table.

Methods

impl<TPeerId, TVal> KBucketsTable<TPeerId, TVal> where
TPeerId: Clone,

pub fn new(local_key: Key<TPeerId>, pending_timeout: Duration) -> Self

Creates a new, empty Kademlia routing table with entries partitioned into buckets as per the Kademlia protocol.

The given pending_timeout specifies the duration after creation of a PendingEntry after which it becomes eligible for insertion into a full bucket, replacing the least-recently (dis)connected node.

pub fn local_key(&self) -> &Key<TPeerId>

Returns the local key.

pub fn entry<'a>(
&'a mut self,
key: &'a Key<TPeerId>
) -> Entry<'a, TPeerId, TVal>

Returns an Entry for the given key, representing the state of the entry in the routing table.

pub fn iter<'a>(
&'a mut self
) -> impl Iterator<Item = EntryRefView<'a, TPeerId, TVal>>

Returns an iterator over all the entries in the routing table.

pub fn buckets<'a>(
&'a mut self
) -> impl Iterator<Item = KBucketRef<'a, TPeerId, TVal>> + 'a

Returns a by-reference iterator over all buckets.

The buckets are ordered by proximity to the local_key, i.e. the first bucket is the closest bucket (containing at most one key).

pub fn take_applied_pending(&mut self) -> Option<AppliedPending<TPeerId, TVal>>

Consumes the next applied pending entry, if any.

When an entry is attempted to be inserted and the respective bucket is full, it may be recorded as pending insertion after a timeout, see InsertResult::Pending.

If the oldest currently disconnected entry in the respective bucket does not change its status until the timeout of pending entry expires, it is evicted and the pending entry inserted instead. These insertions of pending entries happens lazily, whenever the KBucketsTable is accessed, and the corresponding buckets are updated accordingly. The fact that a pending entry was applied is recorded in the KBucketsTable in the form of AppliedPending results, which must be consumed by calling this function.

pub fn closest_keys<'a, T>(
&'a mut self,
target: &'a Key<T>
) -> impl Iterator<Item = Key<TPeerId>> + 'a where
T: Clone,

Returns an iterator over the keys closest to target, ordered by increasing distance.

pub fn closest<'a, T>(
&'a mut self,
target: &'a Key<T>
) -> impl Iterator<Item = EntryView<TPeerId, TVal>> + 'a where
T: Clone,
TVal: Clone,

Returns an iterator over the nodes closest to the target key, ordered by increasing distance.