Struct libp2p_kad::kbucket::Key

pub struct Key<T> { /* fields omitted */ }

A Key in the DHT keyspace with preserved preimage.

Keys in the DHT keyspace identify both the participating nodes, as well as the records stored in the DHT.

Keys have an XOR metric as defined in the Kademlia paper, i.e. the bitwise XOR of the hash digests, interpreted as an integer. See Key::distance.

Implementations

impl<T> Key<T>

pub fn new(preimage: T) -> Key<T> where
T: Borrow<[u8]>,

Constructs a new Key by running the given value through a random oracle.

The preimage of type T is preserved. See Key::preimage and Key::into_preimage.

pub fn preimage(&self) -> &T

Borrows the preimage of the key.

pub fn into_preimage(self) -> T

Converts the key into its preimage.

pub fn distance<U>(&self, other: &U) -> Distance where
U: AsRef<KeyBytes>,

Computes the distance of the keys according to the XOR metric.

pub fn for_distance(&self, d: Distance) -> KeyBytes

Returns the uniquely determined key with the given distance to self.

This implements the following equivalence:

self xor other = distance <==> other = self xor distance