libp2p_kad

Struct Config

pub struct Config { /* private fields */ }

The configuration for the Kademlia behaviour.

The configuration is consumed by Behaviour::new.

Implementations

impl Config

pub fn set_protocol_names(&mut self, names: Vec<StreamProtocol>) -> &mut Self

Sets custom protocol names.

Kademlia nodes only communicate with other nodes using the same protocol name. Using custom name(s) therefore allows to segregate the DHT from others, if that is desired.

More than one protocol name can be supplied. In this case the node will be able to talk to other nodes supporting any of the provided names. Multiple names must be used with caution to avoid network partitioning.

pub fn set_query_timeout(&mut self, timeout: Duration) -> &mut Self

Sets the timeout for a single query.

Note: A single query usually comprises at least as many requests as the replication factor, i.e. this is not a request timeout.

The default is 60 seconds.

pub fn set_replication_factor( &mut self, replication_factor: NonZeroUsize, ) -> &mut Self

Sets the replication factor to use.

The replication factor determines to how many closest peers a record is replicated. The default is K_VALUE.

pub fn set_parallelism(&mut self, parallelism: NonZeroUsize) -> &mut Self

Sets the allowed level of parallelism for iterative queries.

The ฮฑ parameter in the Kademlia paper. The maximum number of peers that an iterative query is allowed to wait for in parallel while iterating towards the closest nodes to a target. Defaults to ALPHA_VALUE.

This only controls the level of parallelism of an iterative query, not the level of parallelism of a query to a fixed set of peers.

When used with Config::disjoint_query_paths it equals the amount of disjoint paths used.

pub fn disjoint_query_paths(&mut self, enabled: bool) -> &mut Self

Require iterative queries to use disjoint paths for increased resiliency in the presence of potentially adversarial nodes.

When enabled the number of disjoint paths used equals the configured parallelism.

See the S/Kademlia paper for more information on the high level design as well as its security improvements.

pub fn set_record_ttl(&mut self, record_ttl: Option<Duration>) -> &mut Self

Sets the TTL for stored records.

The TTL should be significantly longer than the (re-)publication interval, to avoid premature expiration of records. The default is 36 hours.

None means records never expire.

Does not apply to provider records.

pub fn set_record_filtering(&mut self, filtering: StoreInserts) -> &mut Self

Sets whether or not records should be filtered before being stored.

See StoreInserts for the different values. Defaults to StoreInserts::Unfiltered.

pub fn set_replication_interval( &mut self, interval: Option<Duration>, ) -> &mut Self

Sets the (re-)replication interval for stored records.

Periodic replication of stored records ensures that the records are always replicated to the available nodes closest to the key in the context of DHT topology changes (i.e. nodes joining and leaving), thus ensuring persistence until the record expires. Replication does not prolong the regular lifetime of a record (for otherwise it would live forever regardless of the configured TTL). The expiry of a record is only extended through re-publication.

This interval should be significantly shorter than the publication interval, to ensure persistence between re-publications. The default is 1 hour.

None means that stored records are never re-replicated.

Does not apply to provider records.

pub fn set_publication_interval( &mut self, interval: Option<Duration>, ) -> &mut Self

Sets the (re-)publication interval of stored records.

Records persist in the DHT until they expire. By default, published records are re-published in regular intervals for as long as the record exists in the local storage of the original publisher, thereby extending the records lifetime.

This interval should be significantly shorter than the record TTL, to ensure records do not expire prematurely. The default is 24 hours.

None means that stored records are never automatically re-published.

Does not apply to provider records.

pub fn set_provider_record_ttl(&mut self, ttl: Option<Duration>) -> &mut Self

Sets the TTL for provider records.

None means that stored provider records never expire.

Must be significantly larger than the provider publication interval.

pub fn set_provider_publication_interval( &mut self, interval: Option<Duration>, ) -> &mut Self

Sets the interval at which provider records for keys provided by the local node are re-published.

None means that stored provider records are never automatically re-published.

Must be significantly less than the provider record TTL.

pub fn set_connection_idle_timeout(&mut self, duration: Duration) -> &mut Self

๐Deprecated: Set a global idle connection timeout via SwarmBuilder::idle_connection_timeout instead.

Sets the amount of time to keep connections alive when theyโre idle.

pub fn set_max_packet_size(&mut self, size: usize) -> &mut Self

Modifies the maximum allowed size of individual Kademlia packets.

It might be necessary to increase this value if trying to put large records.

pub fn set_kbucket_inserts(&mut self, inserts: BucketInserts) -> &mut Self

Sets the k-bucket insertion strategy for the Kademlia routing table.

pub fn set_caching(&mut self, c: Caching) -> &mut Self

Sets the Caching strategy to use for successful lookups.

The default is Caching::Enabled with a max_peers of 1. Hence, with default settings and a lookup quorum of 1, a successful lookup will result in the record being cached at the closest node to the key that did not return the record, i.e. the standard Kademlia behaviour.