Struct libp2p_kad::Kademlia

pub struct Kademlia<TSubstream> { /* fields omitted */ }

Network behaviour that handles Kademlia.

Methods

impl<TSubstream> Kademlia<TSubstream>

pub fn new(local_peer_id: PeerId) -> Self

Creates a Kademlia.

pub fn without_init(local_peer_id: PeerId) -> Self

Deprecated:
this function is now equivalent to new() and will be removed in the future

Creates a Kademlia.

Contrary to new, doesn't perform the initialization queries that store our local ID into the DHT and fill our buckets.

pub fn add_connected_address(&mut self, peer_id: &PeerId, address: Multiaddr)

Adds a known address for the given PeerId. We are connected to this address.

pub fn add_not_connected_address(
&mut self,
peer_id: &PeerId,
address: Multiaddr
)

Adds a known address for the given PeerId. We are not connected or don't know whether we are connected to this address.

pub fn kbuckets_entries(&self) -> impl Iterator<Item = &PeerId>

Returns an iterator to all the peer IDs in the bucket, without the pending nodes.

pub fn find_node(&mut self, peer_id: PeerId)

Starts an iterative FIND_NODE request.

This will eventually produce an event containing the nodes of the DHT closest to the requested PeerId.

pub fn get_providers(&mut self, target: Multihash)

Starts an iterative GET_PROVIDERS request.

pub fn add_providing(&mut self, key: PeerId)

Register the local node as the provider for the given key.

This will periodically send ADD_PROVIDER messages to the nodes closest to the key. When someone performs a GET_PROVIDERS iterative request on the DHT, our local node will be returned as part of the results.

The actual meaning of providing the value of a key is not defined, and is specific to the value whose key is the hash.

pub fn remove_providing(&mut self, key: &Multihash)

Cancels a registration done with add_providing.

There doesn't exist any "remove provider" message to broadcast on the network, therefore we will still be registered as a provider in the DHT for as long as the timeout doesn't expire.