Enum libp2p_kad::KademliaEvent

pub enum KademliaEvent {
    InboundPutRecordRequest {
        source: PeerId,
        connection: ConnectionId,
        record: Record,
    },
    InboundAddProviderRequest {
        record: ProviderRecord,
    },
    InboundRequestServed {
        request: InboundRequest,
    },
    OutboundQueryCompleted {
        id: QueryId,
        result: QueryResult,
        stats: QueryStats,
    },
    RoutingUpdated {
        peer: PeerId,
        is_new_peer: bool,
        addresses: Addresses,
        bucket_range: (Distance, Distance),
        old_peer: Option<PeerId>,
    },
    UnroutablePeer {
        peer: PeerId,
    },
    RoutablePeer {
        peer: PeerId,
        address: Multiaddr,
    },
    PendingRoutablePeer {
        peer: PeerId,
        address: Multiaddr,
    },
}

The events produced by the Kademlia behaviour.

See NetworkBehaviour::poll.

Variants

InboundPutRecordRequest

Fields

source: PeerId

connection: ConnectionId

record: Record

A peer sent a KademliaHandlerIn::PutRecord request and filtering is enabled.

See KademliaStoreInserts and KademliaConfig::set_record_filtering.

InboundAddProviderRequest

Fields

record: ProviderRecord

A peer sent a KademliaHandlerIn::AddProvider request and filtering KademliaStoreInserts::FilterBoth is enabled.

See KademliaStoreInserts and KademliaConfig::set_record_filtering for details..

InboundRequestServed

Fields

request: InboundRequest

An inbound request has been received and handled.

OutboundQueryCompleted

Fields

id: QueryId

The ID of the query that finished.

result: QueryResult

The result of the query.

stats: QueryStats

Execution statistics from the query.

An outbound query has produced a result.

RoutingUpdated

Fields

peer: PeerId

The ID of the peer that was added or updated.

is_new_peer: bool

Whether this is a new peer and was thus just added to the routing table, or whether it is an existing peer who’s addresses changed.

addresses: Addresses

The full list of known addresses of peer.

bucket_range: (Distance, Distance)

Returns the minimum inclusive and maximum inclusive Distance for the bucket of the peer.

old_peer: Option<PeerId>

The ID of the peer that was evicted from the routing table to make room for the new peer, if any.

The routing table has been updated with a new peer and / or address, thereby possibly evicting another peer.

UnroutablePeer

Fields

peer: PeerId

A peer has connected for whom no listen address is known.

If the peer is to be added to the routing table, a known listen address for the peer must be provided via Kademlia::add_address.

RoutablePeer

Fields

peer: PeerId

address: Multiaddr

A connection to a peer has been established for whom a listen address is known but the peer has not been added to the routing table either because KademliaBucketInserts::Manual is configured or because the corresponding bucket is full.

If the peer is to be included in the routing table, it must must be explicitly added via Kademlia::add_address, possibly after removing another peer.

See Kademlia::kbucket for insight into the contents of the k-bucket of peer.

PendingRoutablePeer

Fields

peer: PeerId

address: Multiaddr

A connection to a peer has been established for whom a listen address is known but the peer is only pending insertion into the routing table if the least-recently disconnected peer is unresponsive, i.e. the peer may not make it into the routing table.

If the peer is to be unconditionally included in the routing table, it should be explicitly added via Kademlia::add_address after removing another peer.

See Kademlia::kbucket for insight into the contents of the k-bucket of peer.