Module libp2p_kad::kbucket

👎Deprecated: The kbucket module will be made private in the future and should not be depended on.

Re-exports

pub use super::bucket::K_VALUE;

Structs

AbsentEntry

An entry that is not present in any bucket.

AppliedPending

The result of applying a pending node to a bucket, possibly replacing an existing node.

Distance

A distance between two keys in the DHT keyspace.

EntryRefView

An immutable by-reference view of a bucket entry.

EntryView

A cloned, immutable view of an entry that is either present in a bucket or pending insertion.

KBucketRef

A reference to a bucket in a KBucketsTable.

KBucketsTable

A KBucketsTable represents a Kademlia routing table.

Key

A Key in the DHT keyspace with preserved preimage.

KeyBytes

The raw bytes of a key in the DHT keyspace.

Node

A Node in a bucket, representing a peer participating in the Kademlia DHT together with an associated value (e.g. contact information).

NodeRefView

An immutable by-reference view of a Node.

PendingEntry

An entry waiting for a slot to be available in a bucket.

PresentEntry

An entry present in a bucket.

Enums

Entry

A reference into a single entry of a KBucketsTable.

InsertResult

The result of inserting an entry into a bucket.

NodeStatus

The status of a node in a bucket.